Lots of new queries
Most are taken from maitenance_schema. But it's not always possible to create a foreign extension. 
- duplicate Indexes
- duplicate FKs
- unused tables
- useless columns (same value in all the table: NULL, blank, or TRUE/FALSE/1/a timestamp...)
- alert on wrap-around
- unindexed foreign keys
- empty tables 
- useless unique constraints on foreign keys
- tables with no PK
- tables of more than 500 tuples without index or PK
- complete report in html format for easier reading

pg_stat_statements new queries (Top 5 only for effectiveness and concise output fore readability)
- queries creating the most IOs (needs track_io_timing=on)
- most time consuming queries overall
- most time consuming in one call
- most severe response jitter (statistics issue)
- most memory consuming
- most temp files 
- most called query 
- complete report for all these queries in html format for readability

The html format is for reporting. There might be a way to make dynamic. Like start.psql. Did not dive into it yet.